requirements.txt
django-crontab==0.7.1
django-crontab
celery
django-celery-beat
redis
